Address Copilot review on #259
Six valid Copilot findings:

1. Circular import — moved reference machinery to its own
   `packages/stagebook/src/schemas/reference.ts`. The walker
   (`validateReferences.ts`) imports from there instead of from
   `./treatment.js`, breaking the cycle. `treatment.ts` re-exports
   the reference symbols so existing barrel imports keep working.
   `nameSchema` moved to a tiny new `schemas/primitives.ts` so
   `reference.ts` doesn't drag in `treatment.ts`.

2. parseDottedReference now re-validates its parsed output via
   `namedReferenceSchema` / `externalReferenceSchema`, so the
   string-shorthand branch enforces the same `nameSchema` regex
   and non-empty path-segment constraints as the structured form.
   The two forms are now exactly equivalent.

3. `getReferenceKeyAndPath` now returns a defensive copy of the
   `path` for both named and external references, so callers can
   safely mutate without leaking into the shared
   `NAMED_SOURCE_DEFAULTS` array or the caller-provided ref.

4. `Element.tsx` Display case now renders the `data-reference`
   attribute via the shared `formatReference()` helper for both
   string and structured input — downstream tooling that scrapes
   the attribute keeps the familiar dotted shape regardless of
   what the treatment file authored or how the host parsed it.

5. `docs/researcher/conditions.md` — replaced the
   "format is `type.name.path`" line with the two grammars (named
   `source.name(.path...)` vs external `source.path...`) so the
   docs match the post-#240 schema.

6. `Display.tsx` `DisplayProps.reference` JSDoc rewritten to
   describe the actual contract (a stringified, dotted-form
   representation routed through `formatReference`) rather than
   implying the raw authored shape passes through unchanged.
